Extract versionProvider.electron
